pyproject.toml
[project]
name = "memsys"
version = "0.1.0"
description = "Memory System with Dynamic and Static Memory Management"
license = {text = "Apache-2.0"}
requires-python = ">=3.12,<3.13"
dependencies = [
    # 核心依赖 - Core Dependencies
    "langgraph>=0.2.6",
    "langgraph-api",
    "langgraph-cli",
    "langgraph-sdk>=0.1.57",
    "langgraph-checkpoint-postgres>=2.0.21",
    "langchain>=0.3.26",
    "langchain-core>=0.1.0",
    "langchain-openai>=0.3.28",
    "langchain-anthropic>=0.3.17",
    "langchain-google-genai>=2.1.8",
    "google-genai>=1.21.0",
    # 数据处理 - Data Processing
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    # 机器学习/AI - Machine Learning/AI
    "sentence-transformers>=2.2.0",
    "torch>=2.6.0",
    "transformers",
    # faiss 版本根据环境选择，见 dependency-groups
    # OpenAI API
    "openai>=1.0.0",
    # 数据验证和配置 - Data Validation & Configuration
    "pydantic[email]>=2.7.1",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.0",
    # Web框架和API - Web Framework & API
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "greenlet>=3.2.0",
    # HTTP客户端和文件处理 - HTTP Client & File Processing
    "aiohttp>=3.8.0",
    "python-multipart>=0.0.6",
    "aiofiles>=24.1.0",
    # 消息队列 - Message Queue
    "aiokafka>=0.11.0",
    # 认证和安全 - Authentication & Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    # 数据库和缓存 - Database & Cache
    "redis>=5.0.0",
    "sqlmodel>=0.0.19",
    "asyncpg>=0.29.0",
    "psycopg[binary,pool]>=3.1.0",
    # MongoDB ODM - MongoDB Object Document Mapper
    "beanie>=1.26.0", # 基于Pydantic的异步MongoDB ODM
    "pymongo>=4.6.0", # MongoDB官方Python驱动（Beanie底层依赖）
    "motor>=3.3.0", # 异步MongoDB驱动（Beanie底层依赖）
    # Elasticsearch
    "elasticsearch-dsl[async]>=8.17.1,<9",
    # Milvus
    "pymilvus~=2.5.0",
    # 分词
    "jieba==0.42.1",
    # 图处理 - Graph Processing
    "igraph>=0.10.0",
    # 工具库 - Utilities
    "tqdm>=4.65.0",
    "arq>=0.26.0",
    "python-snappy>=0.7.3",
    "tzlocal>=5.3.0",
    "rank-bm25>=0.2.2",
    "duckdb>=1.4.0",
    "import-parent-dir>=0.1.1",
    "bs4>=0.0.2",
    "html2text>=2025.4.15",
    "certifi>=2025.10.5",
    "aiolimiter>=1.2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests", "unit_test"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[dependency-groups]
# 开发依赖 - 本地开发时使用
dev = [
    "black>=25.1.0",
    "ipdb>=0.13.13", 
    "isort>=6.0.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.1.0",
    "nest-asyncio",
    "typer",
    "pyinstrument",
    "py-spy",
]

# CPU 版本 faiss - 本地开发和 macOS 使用
cpu = [
    "faiss-cpu>=1.7.0",
]

# GPU 版本 faiss - 生产环境使用
gpu = [
    "faiss-cpu>=1.7.0",
    # "faiss-gpu-cu12>=1.7.0", 0928之前不用
]

# 完整开发环境 - 包含开发工具和 CPU 版本 faiss
dev-full = [
    {include-group = "dev"},
    {include-group = "cpu"},
]

# 生产环境 - 仅包含 GPU 版本 faiss
prod = [
    {include-group = "gpu"},
]

[project.scripts]
bootstrap = "src.bootstrap:main"
web = "src.run:main"
manage = "src.manage:cli"
